% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_min_max_text.R
\name{add_change_text}
\alias{add_change_text}
\title{Find the local min and max values and calculate the moves beetween the local min and maximum values}
\usage{
add_change_text(df, ndays, min_max_text = T)
}
\arguments{
\item{df}{OLHC df, you can get using the get_one_ticker function}

\item{ndays}{Number of days to check for local min or max}
}
\description{
Find the local min and max values and calculate the moves beetween the local min and maximum values
}
