% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_top_n_stock.R
\name{download_top_stock}
\alias{download_top_stock}
\title{Downloading and adding technical indicators to the historical data of a ticker}
\usage{
download_top_stock(location, number_of_stocks = 100, start_date = "1900-01-01")
}
\arguments{
\item{location}{Place where the data should be downloaded, use full paths}

\item{number_of_stocks}{The top n number of stocks that will be downloaded}

\item{start_date}{Start date of the stock data}
}
\description{
Downloading and adding technical indicators to the historical data of a ticker
}
