% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_stock.R
\name{get_one_ticker}
\alias{get_one_ticker}
\title{Downloading and adding technical indicators to the historical data of a ticker}
\usage{
get_one_ticker(
  ticker,
  start_date = "2000-01-01",
  end_date = Sys.Date(),
  mas = c(50, 100, 200)
)
}
\arguments{
\item{ticker}{The ticker of the company to download}

\item{start_date}{The first date of the historical data}

\item{end_date}{The last date of the data}

\item{mas}{List of the simple moving averages to calculate}

\item{ad_local_text}{if you want to calculate the local min max with 20 period}

\item{ad_percent_change}{if you want to calculate the percent change from the first day}
}
\description{
Downloading and adding technical indicators to the historical data of a ticker
}
