% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_plot.R
\name{stock_show_one_plot}
\alias{stock_show_one_plot}
\title{Plot one stock}
\usage{
stock_show_one_plot(
  ticker,
  start_date = (Sys.Date() - 500),
  end_date = Sys.Date(),
  local_point_days = 20,
  fullhd = FALSE
)
}
\arguments{
\item{ticker}{The ticker id of the company}

\item{start_date}{start date of the plot}

\item{end_date}{end date of the plot}

\item{local_point_days}{numbar of days to became a local point}

\item{fullhd}{if you want to save as a full hd png}
}
\description{
Plot one stock
}
